+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2024-11-09
+
+### Added
+
+- Added a link to BoltJamison's Prism crafting guide to the Resources page.
+
 ## 2024-11-08
 
 ### Changed